Updated the controller to filter data in the backend
